completed further part 2
