+++ rgbtohex.js
@@ -1,0 +1,3 @@
+module.exports = function rgbToHex(r, g, b) {
+  return "#" + ((1 << 24) + (r << 16) + (g << 8) + b).toString(16).slice(1);
+}
